tests/wpt/mozilla/tests/mozilla/bluetooth/stopNotifications/stop-after-start-succeeds.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => service.getCharacteristic(heart_rate_measurement.name))
    .then(characteristic => {
        return characteristic.startNotifications()
        .then(start_characteristic => {
            assert_equals(start_characteristic, characteristic);
            return characteristic.stopNotifications()
            .then(stop_characteristic => assert_equals(stop_characteristic, start_characteristic));
        });
    });
}, 'Single stop after start notifications succeeds.');
</script>
